add some function for version 4

+++ blog_web/blogApp/views.py
@@ -1,5 +1,8 @@
 from django.shortcuts import render, redirect, get_object_or_404
+from django.contrib.auth.models import User
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 
 from .models import Post
 
@@ -15,30 +18,38 @@ def index(request):
     return render(request, 'blogApp/index.html', {'posts':posts })
 
 
+@login_required
 def create(request):
     if request.method == 'POST':
+        author = request.user
         title = request.POST['title']
         content = request.POST['content']
-        Post.objects.create(title=title, content=content)
+        Post.objects.create(author=author, title=title, content=content)
         messages.success(request, 'Create Completed.')
         return redirect('blogApp:home')
     return render(request, 'blogApp/create.html')
 
 
+@login_required
 def deletePost(request, post_id):
     post = get_object_or_404(Post, pk=post_id)
     
     if request.method == 'POST':
-        post.delete()
-        messages.success(request, 'Delete sucessful.')
+        author = User.objects.get(pk=post.author)
+        if request.user == author:
+            post.delete()
+            messages.success(request, 'Delete sucessful.')
+        else:
+            messages.error(request, 'You are not authorized to delete this post.')
         return redirect('blogApp:home')
 
     return render(request, 'blogApp/confirm_delete.html', {'post':post})
 
 
+@login_required
 def deleteListPost(request):
     if request.method=='POST':
-        list_post_id = request.POST.getlist('select-post')
+        list_post_id = request.POST.getlist('post-select')
         if list_post_id:
             author = request.user
             Post.objects.filter(id__in=list_post_id, author=author).delete()
@@ -50,6 +61,7 @@ def deleteListPost(request):
     return redirect('blogApp:home')
 
 
+@login_required
 def updatePost(request, post_id):
     post = get_object_or_404(Post, pk=post_id)
     if request.method == 'POST':
